Added functions necessary to get predict working
Takes files from the PromptCC repository written by the original
repository writers. Safe to assume that they had the other repository in
their pythonpath, which I did not want to do.

Also note: the correct model checkpoint to use is located in the
PromptCC repository as well, not the Change-Agent repository